Adds LDAP support
Currently, swauth does not support the possibilty to authenticate a user using an LDAP directory server,
although many enterprises are using LDAP servers for their user accounts.
Organizations which want to use swift with LDAP are obliged to use keystone,
also if they are only interested in a standalone swift installation (without the other openstack components).
This commit adds the possibility of using LDAP to authenticate when a token is going to be generated.

Change-Id: Ieaf8e08ad46e7b78c180e54cfdddd9bc949cabde